handle uni v2 positions as a special case of uni v3 positions
using the full tick range
