+++ packages/subgraph-base/subgraph.yaml
@@ -53,7 +53,7 @@ dataSources:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/factory.ts
+      file: ./src/uniswapV2/factory.ts
       entities:
         - User
       abis:
@@ -74,7 +74,7 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      file: ./src/pair.ts
+      file: ./src/uniswapV2/pair.ts
       entities:
         - AMMPair
       abis:
