+++ packages/subgraph-gnosis-chain/package.json
@@ -6,7 +6,7 @@
     "doall": "rm -rf generated && rm -rf build && graph codegen && graph build",
     "codegen": "graph codegen",
     "build": "graph build",
-    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ cristovaoth/gno-voting-power-gc",
+    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ jfschwarz/gno-voting-power-gc",
     "test": "graph test",
     "create-local": "graph create --node http://localhost:8020/ cristovaoth/gno-voting-power-gc",
     "remove-local": "graph remove --node http://localhost:8020/ cristovaoth/gno-voting-power-gc",
